comments and docstrings are updated.
