+++ examples/conversation_with_RAG_agents/configs/agent_config.json
@@ -16,7 +16,6 @@
                     "module": "llama_index.core",
                     "class": "SimpleDirectoryReader",
                     "init_args": {
-                      "filename_as_id": true,
                       "input_dir": "../../docs/sphinx_doc/en/source/tutorial",
                       "required_exts": [
                         ".md"
@@ -115,7 +114,7 @@
           "similarity_top_k": 3,
           "log_retrieval": true,
           "recent_n_mem": 1,
-          "persist_dir": "../../rag_storage/api_assist",
+          "persist_dir": "./rag_storage/api_assist",
           "repo_base": "../../",
           "file_dir": "../../docs/docstring_html/"
           }
